// www/src/app/request/common/enum/position-history-types.ts
export enum PositionHistoryTypes {
  DOC_REQUEST = 'DOC_REQUEST',
  DOC_POSITION = 'DOC_POSITION',
  DOC_OFFER = 'DOC_OFFER',
  DOC_TECHPROPOSALS = 'DOC_TECHPROPOSALS',
  DOC_MESSAGE = 'DOC_MESSAGE',
  DOC_CONTRACT = 'DOC_CONTRACT',
  DOC_MANUFACTURING = 'DOC_MANUFACTURING',
  REQUEST_ADDED = 'REQUEST_ADDED',
  REQUEST_EDITED = 'REQUEST_EDITED',
  REQUEST_STATUS = 'REQUEST_STATUS',
  POSITION_ADDED = 'POSITION_ADDED',
  POSITION_EDITED = 'POSITION_EDITED',
  POSITION_STATUS = 'POSITION_STATUS',
  OFFER_ADDED = 'OFFER_ADDED',
  OFFER_EDITED = 'OFFER_EDITED',
  OFFER_STATUS = 'OFFER_STATUS',
  WINNER_ADDED = 'WINNER_ADDED',
  WINNER_REMOVED = 'WINNER_REMOVED',
  CONTRACT_ADDED = 'CONTRACT_ADDED',
  CONTRACT_EDITED = 'CONTRACT_EDITED'
}